requirements.txt
########################################
# requirements.txt
########################################
# MCP & LLM
mcp>=1.2.0
openai>=1.40.0
# 이미지/메타데이터
Pillow>=10.3.0
piexif>=1.1.3
# 오디오/플레이리스트
spotipy>=2.24.0
# 데이터 처리
pydantic>=2.8.2
python-dotenv>=1.0.1
requests>=2.31.0